Parse TapestryConfig.
