+++ Package.resolved
@@ -29,6 +29,15 @@
         }
       },
       {
+        "package": "PackageConfig",
+        "repositoryURL": "https://github.com/shibapm/PackageConfig.git",
+        "state": {
+          "branch": null,
+          "revision": "fd0829aac9851434b3d2db0890e27bc489fc973a",
+          "version": "0.12.2"
+        }
+      },
+      {
         "package": "PathKit",
         "repositoryURL": "https://github.com/kylef/PathKit",
         "state": {
@@ -60,7 +69,7 @@
         "repositoryURL": "https://github.com/apple/swift-llbuild.git",
         "state": {
           "branch": "master",
-          "revision": "f2010ce4cd7a6fcb374b86df7aa2364c50d60a3e",
+          "revision": "d45841450f82fca106d6376bc95da031dfb67db1",
           "version": null
         }
       },
@@ -78,7 +87,7 @@
         "repositoryURL": "https://github.com/tuist/tuist.git",
         "state": {
           "branch": "master",
-          "revision": "8ff63bf71cc48e2d38d852d2b0605aef5efe93ee",
+          "revision": "10ab4c7164587b28e167fe03766de61747b469b0",
           "version": null
         }
       },
@@ -87,7 +96,7 @@
         "repositoryURL": "https://github.com/tuist/XcodeProj",
         "state": {
           "branch": "master",
-          "revision": "f9e804ea547e8c5e046b4891b1d875043198a288",
+          "revision": "8b0bfdbf1dc1819f19735be2800a41541c088709",
           "version": null
         }
       }
